consolidate more math utility functions in common/math.h
