+++ src/common/math.h
@@ -20,6 +20,9 @@
 
 #include <math.h>
 #include <stdint.h>
+#ifdef __SSE__
+#include <xmmintrin.h>
+#endif
 
 // work around missing standard math.h symbols
 /** ln(10) */
@@ -31,6 +34,9 @@
 #ifndef M_PI
 #define M_PI   3.14159265358979323846
 #endif /* !M_PI */
+#ifndef M_PI_F
+#define M_PI_F  3.14159265358979324f
+#endif /* !M_PI_F */
 
 
 #define DT_M_PI_F (3.14159265358979324f)
@@ -127,6 +133,177 @@ static inline void mul_mat_vec_2(const float *m, const float *p, float *o)
   o[0] = p[0] * m[0] + p[1] * m[1];
   o[1] = p[0] * m[2] + p[1] * m[3];
 }
+
+static inline float dt_log2f(const float f)
+{
+#ifdef __GLIBC__
+  return log2f(f);
+#else
+  return logf(f) / logf(2.0f);
+#endif
+}
+
+// fast approximation of expf()
+/****** if you change this function, you need to make the same change in data/kernels/{basecurve,basic}.cl ***/
+static inline float dt_fast_expf(const float x)
+{
+  // meant for the range [-100.0f, 0.0f]. largest error ~ -0.06 at 0.0f.
+  // will get _a_lot_ worse for x > 0.0f (9000 at 10.0f)..
+  const int i1 = 0x3f800000u;
+  // e^x, the comment would be 2^x
+  const int i2 = 0x402DF854u; // 0x40000000u;
+  // const int k = CLAMPS(i1 + x * (i2 - i1), 0x0u, 0x7fffffffu);
+  // without max clamping (doesn't work for large x, but is faster):
+  const int k0 = i1 + x * (i2 - i1);
+  union {
+      float f;
+      int k;
+  } u;
+  u.k = k0 > 0 ? k0 : 0;
+  return u.f;
+}
+
+#if defined(__SSE2__)
+#define ALIGNED(a) __attribute__((aligned(a)))
+#define VEC4(a)                                                                                              \
+  {                                                                                                          \
+    (a), (a), (a), (a)                                                                                       \
+  }
+
+/* SSE intrinsics version of dt_fast_expf */
+static const __m128 dt__fone ALIGNED(64) = VEC4(0x3f800000u);
+static const __m128 femo ALIGNED(64) = VEC4(0x00adf880u);
+static inline __m128 dt_fast_expf_sse2(const __m128 x)
+{
+  __m128 f = dt__fone + (x * femo);                 // f(n) = i1 + x(n)*(i2-i1)
+  __m128i i = _mm_cvtps_epi32(f);                   // i(n) = int(f(n))
+  __m128i mask = _mm_srai_epi32(i, 31);             // mask(n) = 0xffffffff if i(n) < 0
+  i = _mm_andnot_si128(mask, i);                    // i(n) = 0 if i(n) < 0
+  return _mm_castsi128_ps(i);                       // return *(float*)&i
+}
+#undef ALIGNED
+#undef VEC4
+
+#endif // __SSE2__
+
+// fast approximation of 2^-x for 0<x<126
+/****** if you change this function, you need to make the same change in data/kernels/{denoiseprofile,nlmeans}.cl ***/
+static inline float dt_fast_mexp2f(const float x)
+{
+  const int i1 = 0x3f800000; // bit representation of 2^0
+  const int i2 = 0x3f000000; // bit representation of 2^-1
+  const int k0 = i1 + (int)(x * (i2 - i1));
+  union {
+    float f;
+    int i;
+  } k;
+  k.i = k0 >= 0x800000 ? k0 : 0;
+  return k.f;
+}
+
+// The below version is incorrect, suffering from reduced precision.
+// It is used by the non-local means code in both nlmeans.c and
+// denoiseprofile.c, and fixing it would cause a change in output.
+static inline float fast_mexp2f(const float x)
+{
+  const float i1 = (float)0x3f800000u; // 2^0
+  const float i2 = (float)0x3f000000u; // 2^-1
+  const float k0 = i1 + x * (i2 - i1);
+  union {
+    float f;
+    int i;
+  } k;
+  k.i = k0 >= (float)0x800000u ? k0 : 0;
+  return k.f;
+}
+
+/** Compute ceil value of a float
+ * @remark Avoid libc ceil for now. Maybe we'll revert to libc later.
+ * @param x Value to ceil
+ * @return ceil value
+ */
+static inline float ceil_fast(float x)
+{
+  if(x <= 0.f)
+  {
+    return (float)(int)x;
+  }
+  else
+  {
+    return -((float)(int)-x) + 1.f;
+  }
+}
+
+#if defined(__SSE2__)
+/** Compute absolute value
+ * @param t Vector of 4 floats
+ * @return Vector of their absolute values
+ */ static inline __m128 _mm_abs_ps(__m128 t)
+{
+  static const uint32_t signmask[4] __attribute__((aligned(64)))
+  = { 0x7fffffff, 0x7fffffff, 0x7fffffff, 0x7fffffff };
+  return _mm_and_ps(*(__m128 *)signmask, t);
+}
+#endif
+
+/** Compute an approximate sine.
+ * This function behaves correctly for the range [-pi pi] only.
+ * It has the following properties:
+ * <ul>
+ *   <li>It has exact values for 0, pi/2, pi, -pi/2, -pi</li>
+ *   <li>It has matching derivatives to sine for these same points</li>
+ *   <li>Its relative error margin is <= 1% iirc</li>
+ *   <li>It computational cost is 5 mults + 3 adds + 2 abs</li>
+ * </ul>
+ * @param t Radian parameter
+ * @return guess what
+ */
+static inline float sinf_fast(float t)
+{
+  /***** if you change this function, you must also change the copy in data/kernels/basic.cl *****/
+  static const float a = 4 / (M_PI * M_PI);
+  static const float p = 0.225f;
+
+  t = a * t * (M_PI_F - fabsf(t));
+
+  return t * (p * (fabsf(t) - 1) + 1);
+}
+
+#if defined(__SSE2__)
+/** Compute an approximate sine (SSE version, four sines a call).
+ * This function behaves correctly for the range [-pi pi] only.
+ * It has the following properties:
+ * <ul>
+ *   <li>It has exact values for 0, pi/2, pi, -pi/2, -pi</li>
+ *   <li>It has matching derivatives to sine for these same points</li>
+ *   <li>Its relative error margin is <= 1% iirc</li>
+ *   <li>It computational cost is 5 mults + 3 adds + 2 abs</li>
+ * </ul>
+ * @param t Radian parameter
+ * @return guess what
+ */
+static inline __m128 sinf_fast_sse(__m128 t)
+{
+  static const __m128 a
+      = { 4.f / (M_PI * M_PI), 4.f / (M_PI * M_PI), 4.f / (M_PI * M_PI), 4.f / (M_PI * M_PI) };
+  static const __m128 p = { 0.225f, 0.225f, 0.225f, 0.225f };
+  static const __m128 pi = { M_PI, M_PI, M_PI, M_PI };
+
+  // m4 = a*t*(M_PI - fabsf(t));
+  const __m128 m1 = _mm_abs_ps(t);
+  const __m128 m2 = _mm_sub_ps(pi, m1);
+  const __m128 m3 = _mm_mul_ps(t, m2);
+  const __m128 m4 = _mm_mul_ps(a, m3);
+
+  // p*(m4*fabsf(m4) - m4) + m4;
+  const __m128 n1 = _mm_abs_ps(m4);
+  const __m128 n2 = _mm_mul_ps(m4, n1);
+  const __m128 n3 = _mm_sub_ps(n2, m4);
+  const __m128 n4 = _mm_mul_ps(p, n3);
+
+  return _mm_add_ps(n4, m4);
+}
+#endif
 
 
 // modelines: These editor modelines have been set for all relevant files by tools/update_modelines.sh
